Refactor: Centralisation des types (typography, cta, profile) et suppression des exports redondants
