+++ src/components/common/cta/index.ts
@@ -1,2 +1,1 @@
 export { default as CallToAction } from './CallToAction';
-export type { CallToActionProps } from './CallToAction';
