+++ src/components/common/typography/index.ts
@@ -1,4 +1,2 @@
 export { default as Title } from './Title';
 export { default as Subtitle } from './Subtitle';
-export type { TitleProps } from './Title';
-export type { SubtitleProps } from './Subtitle';
